setup.py
#!/usr/bin/env python3
"""
LocalLLM Setup Script

Installs LocalLLM and all dependencies for local AI chat.
"""

import os
import sys
import subprocess
import platform
import urllib.request
import zipfile
import tarfile
from pathlib import Path
import json

def print_banner():
    """Print installation banner"""
    print("""
🤖 LocalLLM Installer
=====================

Installing your personal AI assistant...
• 100% Local & Private
• No Internet Required After Setup
• Works on Any Computer

""")

def check_python_version():
    """Check if Python version is compatible"""
    print("🐍 Checking Python version...")
    
    if sys.version_info < (3, 8):
        print("❌ Python 3.8+ required")
        print(f"   Current version: {sys.version}")
        print("   Please upgrade Python and try again")
        sys.exit(1)
    
    print(f"✅ Python {sys.version.split()[0]} detected")

def detect_system():
    """Detect operating system and architecture"""
    print("🖥️ Detecting system...")
    
    system = platform.system().lower()
    machine = platform.machine().lower()
    
    # Normalize architecture names
    if machine in ['x86_64', 'amd64']:
        arch = 'x64'
    elif machine in ['aarch64', 'arm64']:
        arch = 'arm64'
    elif machine.startswith('arm'):
        arch = 'arm'
    else:
        arch = 'x64'  # Default fallback
    
    print(f"✅ System: {system.title()} {arch}")
    return system, arch

def install_dependencies():
    """Install Python dependencies"""
    print("📦 Installing dependencies...")
    
    # Core dependencies
    dependencies = [
        "torch>=2.0.0",
        "transformers>=4.30.0",
        "tokenizers>=0.13.0",
        "numpy>=1.21.0",
        "pyyaml>=6.0",
        "tqdm>=4.64.0",
        "requests>=2.28.0",
    ]
    
    # Optional GUI dependencies
    gui_dependencies = [
        "tkinter",  # Usually comes with Python
        "gradio>=3.35.0",  # For web UI
    ]
    
    # Install core dependencies
    for dep in dependencies:
        try:
            print(f"   Installing {dep.split('>=')[0]}...")
            subprocess.check_call([
                sys.executable, "-m", "pip", "install", dep, "--quiet"
            ])
        except subprocess.CalledProcessError as e:
            print(f"⚠️ Warning: Failed to install {dep}")
    
    # Try to install GUI dependencies
    for dep in gui_dependencies:
        try:
            subprocess.check_call([
                sys.executable, "-m", "pip", "install", dep, "--quiet"
            ])
        except subprocess.CalledProcessError:
            pass  # GUI dependencies are optional
    
    print("✅ Dependencies installed")

def download_model():
    """Download the pre-trained model"""
    print("🧠 Downloading AI model...")
    
    model_dir = Path("models")
    model_dir.mkdir(exist_ok=True)
    
    # For now, we'll create a placeholder
    # In a real implementation, you'd download from your model hosting
    model_info = {
        "name": "LocalLLM-Small",
        "version": "1.0.0",
        "size": "1.2GB",
        "description": "Lightweight model optimized for local inference",
        "vocab_size": 32000,
        "hidden_size": 768,
        "num_layers": 12,
        "num_heads": 12
    }
    
    model_config_path = model_dir / "config.json"
    with open(model_config_path, 'w') as f:
        json.dump(model_info, f, indent=2)
    
    # Create placeholder model file
    model_path = model_dir / "model.bin"
    if not model_path.exists():
        print("   Creating model placeholder...")
        # In real implementation, download actual model weights
        with open(model_path, 'wb') as f:
            f.write(b"# LocalLLM Model Placeholder\n")
    
    print("✅ Model ready")

def create_config():
    """Create default configuration"""
    print("⚙️ Creating configuration...")
    
    config = {
        "model": {
            "path": "models/model.bin",
            "config": "models/config.json",
            "max_tokens": 2048,
            "temperature": 0.7,
            "top_p": 0.9
        },
        "performance": {
            "use_gpu": "auto",
            "threads": "auto",
            "memory_limit": "4GB",
            "batch_size": 1
        },
        "interface": {
            "theme": "dark",
            "save_history": True,
            "history_file": "chat_history.json",
            "max_history": 1000
        },
        "api": {
            "host": "localhost",
            "port": 8080,
            "enable_cors": True
        }
    }
    
    with open("config.yaml", 'w') as f:
        import yaml
        yaml.dump(config, f, default_flow_style=False, indent=2)
    
    print("✅ Configuration created")

def create_launcher_scripts():
    """Create launcher scripts for different platforms"""
    print("🚀 Creating launcher scripts...")
    
    system, arch = detect_system()
    
    # Chat launcher
    if system == "windows":
        chat_script = """@echo off
echo Starting LocalLLM Chat...
python chat.py %*
pause
"""
        with open("chat.bat", 'w') as f:
            f.write(chat_script)
    else:
        chat_script = """#!/bin/bash
echo "Starting LocalLLM Chat..."
python3 chat.py "$@"
"""
        with open("chat.sh", 'w') as f:
            f.write(chat_script)
        os.chmod("chat.sh", 0o755)
    
    # GUI launcher
    if system == "windows":
        gui_script = """@echo off
echo Starting LocalLLM GUI...
python gui_chat.py
"""
        with open("gui.bat", 'w') as f:
            f.write(gui_script)
    else:
        gui_script = """#!/bin/bash
echo "Starting LocalLLM GUI..."
python3 gui_chat.py
"""
        with open("gui.sh", 'w') as f:
            f.write(gui_script)
        os.chmod("gui.sh", 0o755)
    
    print("✅ Launcher scripts created")

def run_tests():
    """Run basic functionality tests"""
    print("🧪 Running tests...")
    
    try:
        # Test imports
        import torch
        import transformers
        print("✅ Core libraries working")
        
        # Test model loading (placeholder)
        if Path("models/config.json").exists():
            print("✅ Model files accessible")
        
        print("✅ All tests passed")
        return True
        
    except Exception as e:
        print(f"⚠️ Test failed: {e}")
        return False

def print_success():
    """Print success message with usage instructions"""
    system, arch = detect_system()
    
    print("""
🎉 LocalLLM Installation Complete!

🚀 Quick Start:
""")
    
    if system == "windows":
        print("""   • Double-click 'chat.bat' to start chatting
   • Double-click 'gui.bat' for graphical interface
   • Or run: python chat.py""")
    else:
        print("""   • Run: ./chat.sh (or python3 chat.py)
   • GUI: ./gui.sh (or python3 gui_chat.py)
   • API: python3 api_server.py""")
    
    print("""
📖 Documentation:
   • README.md - Full documentation
   • config.yaml - Customize settings
   • examples/ - Usage examples

🔒 Privacy: Your AI runs 100% locally!
💬 Support: Check README.md for help

Happy chatting! 🤖
""")

def main():
    """Main installation function"""
    try:
        print_banner()
        
        # System checks
        check_python_version()
        system, arch = detect_system()
        
        # Installation steps
        install_dependencies()
        download_model()
        create_config()
        create_launcher_scripts()
        
        # Verification
        if run_tests():
            print_success()
        else:
            print("⚠️ Installation completed with warnings")
            print("   Check README.md for troubleshooting")
        
        return 0
        
    except KeyboardInterrupt:
        print("\n❌ Installation cancelled by user")
        return 1
        
    except Exception as e:
        print(f"\n❌ Installation failed: {e}")
        print("   Please check the error and try again")
        return 1

if __name__ == "__main__":
    exit_code = main()
    sys.exit(exit_code)
